Updated TODO.

+++ src/main/java/org/halvors/quantum/common/tile/reactor/fusion/TilePlasma.java
@@ -11,7 +11,7 @@ import org.halvors.quantum.lib.thermal.ThermalGrid;
 
 public class TilePlasma extends TileEntity {
     public static int plasmaMaxTemperature = 1000000;
-    private float temperature = plasmaMaxTemperature;
+    private int temperature = plasmaMaxTemperature;
 
     @Override
     public void updateEntity() {
@@ -38,7 +38,7 @@ public class TilePlasma extends TileEntity {
                         TileEntity tileEntity = position.getTileEntity(worldObj);
 
                         if (!(tileEntity instanceof TilePlasma)) {
-                            MinecraftForge.EVENT_BUS.post(new PlasmaEvent.PlasmaSpawnEvent(worldObj, position.intX(), position.intY(), position.intZ(), (int) temperature));
+                            MinecraftForge.EVENT_BUS.post(new PlasmaEvent.PlasmaSpawnEvent(worldObj, position.intX(), position.intY(), position.intZ(), temperature));
                         }
                     }
                 }
